refactor: move types into their own directory
