+++ src/selectors/people.ts
@@ -1,18 +1,6 @@
 // Selectors for friends and other people
 
-export interface FriendRights {
-  canSeeOnline: boolean
-  canSeeOnMap: boolean
-  canModifyObjects: boolean
-}
-
-export interface Friend {
-  id: string
-  // from me to friend
-  rightsGiven: FriendRights
-  // Friend has given me rights
-  rightsHas: FriendRights
-}
+import { Friend } from '../types/people'
 
 export const getFriends = (state: any): Friend[] => state.friends
