+++ src/selectors/viewer.ts
@@ -1,34 +1,7 @@
 // Selectors for viewer:
 // Account and other general state
 
-export interface HoodieTimeObject {
-  createdAt: string
-  updatedAt?: string
-  deletedAt?: string
-}
-
-export interface AvatarData {
-  name: string
-  grid: string
-  avatarIdentifier: string
-  dataSaveId: string
-}
-
-export interface SavedAvatarData extends AvatarData {
-  _id: string
-  _rev: string
-  _deleted?: boolean
-  hoodie: HoodieTimeObject
-}
-
-export interface Grid {
-  _id: string
-  _rev: string
-  _deleted?: boolean
-  hoodie: HoodieTimeObject
-  name: string
-  loginURL: string
-}
+import { AvatarData, SavedAvatarData, Grid } from '../types/viewer'
 
 export const getIsSignedIn = (state: any): boolean => state.account.loggedIn
